Reorganize vote extensions mod
