+++ apps/src/lib/node/ledger/protocol/mod.rs
@@ -14,10 +14,10 @@ use namada::ledger::storage::{DBIter, Storage, StorageHasher, DB};
 use namada::ledger::treasury::TreasuryVp;
 use namada::proto::{self, Tx};
 use namada::types::address::{Address, InternalAddress};
-use namada::types::ethereum_events::vote_extensions::EthEventsVextDigest;
 use namada::types::storage;
 use namada::types::transaction::protocol::{ProtocolTx, ProtocolTxType};
 use namada::types::transaction::{DecryptedTx, TxResult, TxType, VpsResult};
+use namada::types::vote_extensions::ethereum_events::EthEventsVextDigest;
 use namada::vm::wasm::{TxCache, VpCache};
 use namada::vm::{self, wasm, WasmCacheAccess};
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
