+++ shared/src/types/mod.rs
@@ -17,3 +17,4 @@ pub mod time;
 pub mod token;
 pub mod transaction;
 pub mod validity_predicate;
+pub mod vote_extensions;
