+++ shared/src/types/vote_extensions.rs
@@ -1,0 +1,3 @@
+//! This module contains types necessary for processing vote extensions.
+
+pub mod ethereum_events;
